Changed CName & NodeRef hashing

+++ WolvenKit.Core/FNV1A/FNV1A64HashAlgorithm.cs
@@ -78,7 +78,7 @@ namespace WolvenKit.Common.FNV1A
             {
                 unchecked
                 {
-                    hash = (hash ^ (uint)b) * FnvHashPrime;
+                    hash = (hash ^ (ulong)b) * FnvHashPrime;
                 }
             }
